requirements-dev.txt
-r requirements.txt
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
coverage>=7.2.7
pre-commit>=3.3.2
types-requests>=2.31.0
types-python-dateutil>=2.8.19
pytest-xdist>=3.3.1 